requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Bcrypt==1.0.1
itsdangerous==2.1.2
mysql-connector-python==8.3.0
Werkzeug==3.0.1
